Move overflow test logic into its own header
- is_overflow.h - rather confusing as it doesn't contain a definition with
  the name (yet)
